WGPUCompute

Documentation for WGPUCompute.

WGPUCompute.cast

WGPUCompute.cast_kernel

WGPUCompute.cast — Method

cast(S::DataType, x::WgpuArray{T, N}) where {T, N}

This is a wrapper function for cast_kernel kernel function. This is meant for users to cast from regular julia functions.

WGPUCompute.cast_kernel — Method

cast_kernel(x::WgpuArray{T, N}, out::WgpuArray{S, N}) where {T, S, N}

This is a compute kernel which casts the x array of eltype T to eltype S. Users are not supposed to use this function call from julia. This instead needs to be wrapped with an additional function which uses @wgpukernel macro call to convert the julia function definition to a equivalent WGPU kernel function.