WGPUCompute

Documentation for WGPUCompute.

WGPUCompute.return_type

WGPUCompute.wgpuconvert

WGPUCompute.wgpufunction

WGPUCompute.return_type — Method

SPIRV.return_type(f, tt) -> r::Type

Return a type r such that f(args...)::r where args::tt.

WGPUCompute.wgpuconvert — Function

wgpuconvert(x, [cce])

This function is called for every argument to be passed to a kernel, allowing it to be converted to a GPU-friendly format. By default, the function does nothing and returns the input object x as-is.

Do not add methods to this function, but instead extend the underlying Adapt.jl package and register methods for the the Metal.Adaptor type.

WGPUCompute.wgpufunction — Method

wgpufunction(f, tt=Tuple{}; kwargs...)

Low-level interface to compile a function invocation for the currently-active GPU, returning a callable kernel object. For a higher-level interface, use @metal.

The output of this function is automatically cached, i.e. you can simply call wgpufunction in a hot path without degrading performance. New code will be generated automatically when the function changes, or when different types or keyword arguments are provided.